pyproject.toml
[project]
name = "eclock-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.0.5",
    "fastapi>=0.119.0",
    "loguru>=0.7.3",
    "openai>=1.107.3",
    "pydantic-settings>=2.10.1",
    "pyinstaller>=6.16.0",
    "pypdf2>=3.0.1",
    "redis>=5.0.0",
    "requests>=2.32.5",
    "tomli>=2.2.1",
    "uvicorn>=0.38.0",
]
